major refactoring and tests
