+++ exampleapps/README.md
@@ -1,0 +1,2 @@
+# Exampleapps
+This is used primary for debugging. Setup your config.json in the binary folder before debug connecting to Home Assistant.
