+++ src/DaemonRunner/DaemonRunner/Service/App/CodeManager.cs
@@ -1,23 +1,19 @@
 using System;
-using System.CodeDom.Compiler;
 using System.Collections.Generic;
+using System.Dynamic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.Loader;
 using System.Text;
+using System.Threading.Tasks;
 using JoySoftware.HomeAssistant.NetDaemon.Common;
 using JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.Config;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.Text;
-// using Microsoft.CodeAnalysis;
-// using Microsoft.CodeAnalysis.CSharp.Scripting;
-// using Microsoft.CodeAnalysis.Emit;
-// using Microsoft.CodeAnalysis.Scripting;
-// using Microsoft.CodeAnalysis.CSharp;
-// using Microsoft.CodeAnalysis.Text;
+using Microsoft.Extensions.Logging;
 
 [assembly: InternalsVisibleTo("NetDaemon.Daemon.Tests")]
 namespace JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.App
@@ -28,11 +24,146 @@ namespace JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.App
         {
         }
 
-        protected override Assembly Load(AssemblyName name)
+        protected override Assembly? Load(AssemblyName _) => null;
+    }
+
+    public static class DaemonAppExtensions
+    {
+        public static void HandleAttributeInitialization(this INetDaemonApp netDaemonApp, INetDaemon _daemon)
         {
-            return null;
+            var netDaemonAppType = netDaemonApp.GetType();
+            foreach (var method in netDaemonAppType.GetMethods())
+            {
+                foreach (var attr in method.GetCustomAttributes(false))
+                {
+                    switch (attr)
+                    {
+                        case HomeAssistantServiceCallAttribute hasstServiceCallAttribute:
+                            HandleServiceCallAttribute(_daemon, netDaemonApp, method);
+
+                            break;
+                        case HomeAssistantStateChangedAttribute hassStateChangedAttribute:
+
+
+
+                            break;
+                    }
+                }
+            }
+        }
+
+        private static void HandleStateChangedAttribute(
+            INetDaemon _daemon,
+            HomeAssistantStateChangedAttribute hassStateChangedAttribute,
+            INetDaemonApp netDaemonApp,
+            MethodInfo method
+            )
+        {
+            var (signatureOk, err) = CheckIfStateChangedSignatureIsOk(method);
+
+            if (!signatureOk)
+            {
+                _daemon.Logger.LogWarning(err);
+                return;
+            }
+
+            _daemon.ListenState(hassStateChangedAttribute.EntityId,
+            async (entityId, to, from) =>
+            {
+                try
+                {
+                    if (hassStateChangedAttribute.To != null)
+                        if ((dynamic)hassStateChangedAttribute.To != to?.State)
+                            return;
+
+                    if (hassStateChangedAttribute.From != null)
+                        if ((dynamic)hassStateChangedAttribute.From != from?.State)
+                            return;
+
+                    // If we don´t accept all changes in the state change
+                    // and we do not have a state change so return
+                    if (to?.State == from?.State && !hassStateChangedAttribute.AllChanges)
+                        return;
+
+                    await method.InvokeAsync(netDaemonApp, entityId, to, from);
+                }
+                catch (Exception e)
+                {
+                    _daemon.Logger.LogError(e, "Failed to invoke the ServiceCall funcition");
+                }
+
+            });
+        }
+
+        private static void HandleServiceCallAttribute(INetDaemon _daemon, INetDaemonApp netDaemonApp, MethodInfo method)
+        {
+            var (signatureOk, err) = CheckIfServiceCallSignatureIsOk(method);
+            if (!signatureOk)
+            {
+                _daemon.Logger.LogWarning(err);
+                return;
+            }
+
+            dynamic serviceData = new FluentExpandoObject();
+            serviceData.method = method.Name;
+            serviceData.@class = netDaemonApp.GetType().Name;
+            _daemon.CallService("netdaemon", "register_service", serviceData);
+
+            _daemon.ListenServiceCall("netdaemon", $"{serviceData.@class}_{serviceData.method}",
+                async (data) =>
+                {
+                    try
+                    {
+                        var expObject = data as ExpandoObject;
+                        await method.InvokeAsync(netDaemonApp, expObject);
+                    }
+                    catch (Exception e)
+                    {
+                        _daemon.Logger.LogError(e, "Failed to invoke the ServiceCall function");
+                    }
+                });
+        }
+        private static (bool, string) CheckIfServiceCallSignatureIsOk(MethodInfo method)
+        {
+            if (method.ReturnType != typeof(Task))
+                return (false, $"{method.Name} has not correct return type, expected Task");
+
+            var parameters = method.GetParameters();
+
+            if (parameters == null || (parameters != null && parameters.Length != 1))
+                return (false, $"{method.Name} has not correct number of parameters");
+
+            var dynParam = parameters![0];
+            if (dynParam.CustomAttributes.Count() == 1 &&
+                dynParam.CustomAttributes.First().AttributeType == typeof(DynamicAttribute))
+                return (true, string.Empty);
+
+            return (false, $"{method.Name} is not correct signature");
+        }
+
+        private static (bool, string) CheckIfStateChangedSignatureIsOk(MethodInfo method)
+        {
+            if (method.ReturnType != typeof(Task))
+                return (false, $"{method.Name} has not correct return type, expected Task");
+
+            var parameters = method.GetParameters();
+
+            if (parameters == null || (parameters != null && parameters.Length != 3))
+                return (false, $"{method.Name} has not correct number of parameters");
+
+            if (parameters![0].ParameterType != typeof(string))
+                return (false, $"{method.Name} first parameter exepected to be string for entityId");
+
+            if (parameters![1].ParameterType != typeof(EntityState))
+                return (false, $"{method.Name} second parameter exepected to be EntityState for toState");
+
+            if (parameters![2].ParameterType != typeof(EntityState))
+                return (false, $"{method.Name} first parameter exepected to be EntityState for fromState");
+
+            return (true, string.Empty);
         }
     }
+
     public class CodeManager
     {
         // private readonly ScriptOptions _scriptOptions;
@@ -43,24 +174,33 @@ namespace JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.App
         {
             _codeFolder = codeFolder;
             _loadedDaemonApps = new List<Type>(100);
-            // _scriptOptions = LoadAssembliesAndStandardImports();
 
             if (!string.IsNullOrEmpty(_codeFolder))
                 LoadAllCodeToLoadContext();
         }
 
-        public IEnumerable<INetDaemonApp> InstanceAndInitApplications(INetDaemonHost host)
+        public IEnumerable<Type> DaemonAppTypes => _loadedDaemonApps;
+
+        public IEnumerable<INetDaemonApp> InstanceAndInitApplications(INetDaemon host)
         {
+            _ = (host as INetDaemon) ?? throw new ArgumentNullException(nameof(host));
+
             var result = new List<INetDaemonApp>();
             foreach (string file in Directory.EnumerateFiles(_codeFolder, "*.yaml", SearchOption.AllDirectories))
             {
                 var appInstance = DaemonAppTypes.InstanceFromYamlConfig(File.OpenText(file));
+
                 if (appInstance != null)
+                {
                     result.Add(appInstance);
-            }            
+                    appInstance.StartUpAsync(host);
+                    appInstance.InitializeAsync();
+                    host!.Logger.LogInformation($"Successfully loaded app {appInstance.GetType().Name}");
+                }
+
+            }
             return result;
         }
-
 
         private void LoadAllCodeToLoadContext()
         {
@@ -74,7 +214,6 @@ namespace JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.App
                     var sourceText = SourceText.From(File.ReadAllText(csFile));
                     var syntaxTree = SyntaxFactory.ParseSyntaxTree(sourceText);
                     syntaxTrees.Add(syntaxTree);
-
                 }
 
                 var metaDataReference = new List<MetadataReference>(10)
@@ -83,7 +222,6 @@ namespace JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.App
                     MetadataReference.CreateFromFile(typeof(Console).Assembly.Location),
                     MetadataReference.CreateFromFile(typeof(System.Runtime.AssemblyTargetedPatchBandAttribute).Assembly.Location),
                     MetadataReference.CreateFromFile(typeof(Microsoft.CSharp.RuntimeBinder.CSharpArgumentInfo).Assembly.Location),
-
                 };
 
                 var assembliesFromCurrentAppDomain = AppDomain.CurrentDomain.GetAssemblies();
@@ -128,15 +266,9 @@ namespace JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.App
                     }
                     var err = msg.ToString();
                     System.Console.WriteLine(err);
-
                 }
-
-
             }
-
         }
-
-        public IEnumerable<Type> DaemonAppTypes => _loadedDaemonApps;
 
         public static IEnumerable<string> GetCsFiles(string configFixturePath)
         {
@@ -145,76 +277,3 @@ namespace JoySoftware.HomeAssistant.NetDaemon.DaemonRunner.Service.App
 
     }
 }
-
-
-
-// private void LoadAllCodeToLoadContext2()
-// {
-//     var stream = new MemoryStream();
-//     var alc = new CollectibleAssemblyLoadContext();
-
-//     // EmitResult emitResult;
-//     // Script<object>? script = null;
-//     foreach (var csFile in GetCsFiles(_codeFolder))
-//     {
-//         var script = CSharpScript.Create(File.ReadAllText(csFile), _scriptOptions);
-//         var compilation = script.GetCompilation();
-//         var emitResult = compilation.Emit(stream);
-
-//         stream.Seek(0, SeekOrigin.Begin);
-//         if (emitResult.Success) //
-//         {
-
-//             var asm = alc.LoadFromStream(stream);
-//             var assemblyAppTypes = asm.GetTypes().Where(type => type.IsClass && type.IsSubclassOf(typeof(NetDaemonApp)));
-//             foreach (var app in assemblyAppTypes)
-//             {
-//                 _loadedDaemonApps.Add(app);
-//             }
-
-
-//         }
-//         else
-//         {
-//             var msg = new StringBuilder();
-//             msg.AppendLine($"Compiler error!");
-
-//             foreach (var emitResultDiagnostic in emitResult.Diagnostics)
-//             {
-//                 if (emitResultDiagnostic.Severity == DiagnosticSeverity.Error)
-//                 {
-//                     msg.AppendLine(emitResultDiagnostic.ToString());
-//                 }
-//             }
-//             var err = msg.ToString();
-//             System.Console.WriteLine(err);
-
-//         }
-//     }
-//     alc.Unload();
-//     GC.Collect();
-//     GC.WaitForPendingFinalizers();
-
-
-
-
-// }
-
-// private ScriptOptions LoadAssembliesAndStandardImports()
-// {
-//     // Load all assemblies from the current running daemon
-//     var options = ScriptOptions.Default;
-
-//     var assembliesFromCurrentAppDomain = AppDomain.CurrentDomain.GetAssemblies();
-//     foreach (var assembly in assembliesFromCurrentAppDomain)
-//     {
-//         // _logger.LogInformation(assembly.FullName);
-//         if (assembly.FullName != null && assembly.FullName.StartsWith("NetDaemon"))
-//             options = options.AddReferences(assembly);
-//     }
-
-//     // Add the standard imports
-//     options = options.AddImports("System");
-//     options = options.AddImports("JoySoftware.HomeAssistant.NetDaemon.Common");
-//     return options;
-// }
