--- tests/NetDaemon.Daemon.Tests/DaemonRunner/App/AssemblyDaemonApp.yaml
@@ -1,3 +1,0 @@
-assembly_app:
-    class: AssmeblyDaemonApp
-
